Moved treewalk code to its own module
